Extract BaseRenderer from ScriptRenderer
